pyproject.toml
[tool.poetry]
name = "hexlet_code"
version = "0.1.5"
description = ""
authors = ["Егор <oegor8194@gmail.com>"]
packages = [
	{ include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
flake8 = "^3.8.4"
pytest = "^6.2.1"
pytest-cov = "^2.10.1"
PyYAML = "^5.3.1"
colorama = "^0.4.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
